% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anesrake.R
\name{anesrake}
\alias{anesrake}
\title{Backwards compatibility function call syntax for anesrake}
\usage{
anesrake(inputter, dataframe, caseid = NULL, weightvec = NULL, cap = 5,
  verbose = FALSE, maxit = 1000, type = "pctlim", pctlim = 5,
  nlim = 5, filter = NULL, choosemethod = "total", iterate = NULL,
  convcrit = 0.01, force1 = NULL, center.baseweights = NULL)
}
\arguments{
\item{inputter}{Mapped to \code{\link{harvest}}'s \code{target} argument}

\item{dataframe}{Mapped to \code{\link{harvest}}'s \code{data} argument}

\item{caseid}{Ignored}

\item{weightvec}{Mapped to \code{\link{harvest}}'s \code{start_weights}
argument}

\item{cap}{Mapped to \code{\link{harvest}}'s \code{max_weight} argument}

\item{verbose}{Mapped to \code{\link{harvest}}'s \code{verbose} argument}

\item{maxit}{Mapped to \code{\link{harvest}}'s \code{max_iterations}
argument}

\item{type}{Mapped to \code{\link{harvest}}'s \code{select_function}
argument}

\item{pctlim}{Mapped to \code{\link{harvest}}'s
\code{select_params[["pct"]]} argument}

\item{nlim}{Mapped to \code{\link{harvest}}'s
\code{select_params[["number"]]} argument}

\item{filter}{Ignored}

\item{choosemethod}{Mapped to \code{\link{harvest}}'s \code{error_function}
argument}

\item{iterate}{Ignored, \code{\link{harvest}}'s functionality is equivalent
to \code{iterate = TRUE}}

\item{convcrit}{Mapped to \code{\link{harvest}}'s
\code{convergence[["pct"]]} argument}

\item{force1}{Ignored, \code{\link{harvest}}'s functionality is equivalent
to \code{force1 = TRUE}}

\item{center.baseweights}{Ignored, \code{\link{harvest}}'s functionality is
equivalent to \code{iterate = TRUE}}
}
\value{
A list containing one item, a named vector weightvec which contains
  the weights generated by \code{\link{harvest}}. Other elements of
  \code{\link[anesrake]{anesrake}} returns are not supported.
}
\description{
This function exists to provide a backwards compatible function syntax for
\code{\link[anesrake]{anesrake}} and should not be called for any other
reason. Rewriting the request as a \code{\link{harvest}} call is preferable.
}
\details{
The arguments \code{iterate}, \code{force1}, and \code{center.baseweights}
are not supported. \code{iterate} and \code{center.baseweights} are forced to
be TRUE. \code{force1} is ignored and if any target variables do not sum to 1,
an error will occur.
}
\examples{
\dontrun{
# Load autumn: this will mask anesrake's anesrake function with autumn's
library(autumn)

# Argument order and names are the same as anesrake; but caseid is no longer
# required and the increased flexibility for data and target arguments from
# `harvest()` are supported.
anesrake(inputter = ns_target, dataframe = respondent_data)

# Most anesrake arguments supported
anesrake(ns_target, respondent_data, type = "nlim", choosemethod = "max")
}
}
