% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_mm.R
\name{auto_mm}
\alias{auto_mm}
\title{Automate Many Model Analysis}
\usage{
auto_mm(dataframe, model, split_vars, min_sample_size = 30)
}
\arguments{
\item{dataframe}{Dataframe to operate on}

\item{min_sample_size}{Minimum sample size of the groups}

\item{model_function}{Model function to test with}

\item{vars}{Variables to cut data with}

\item{model_metrics}{Metrics to pull from model}
}
\description{
Automated many-model approach using functional programming.
}
