% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rake.R
\name{clamp_weights_top}
\alias{clamp_weights_top}
\title{Clamp weights to a maximum weight}
\usage{
clamp_weights_top(weights, clamp)
}
\arguments{
\item{weights}{Current vector of weights}

\item{clamp}{A scalar maximum weight}
}
\value{
A vector of clamped weights
}
\description{
Clamp weights to a maximum weight
}
\keyword{internal}
