% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/current_miss.R
\name{current_miss_linear}
\alias{current_miss_linear}
\alias{current_miss_squared}
\alias{current_miss_mean}
\alias{current_miss_max}
\alias{current_miss_max_squared}
\alias{current_miss_mean_squared}
\title{Error Function}
\usage{
current_miss_linear(truth, current)

current_miss_squared(truth, current)

current_miss_mean(truth, current)

current_miss_max(truth, current)

current_miss_max_squared(truth, current)

current_miss_mean_squared(truth, current)
}
\arguments{
\item{truth}{A vector containing the true proportions of the target variable}

\item{current}{A vector containing the weighted proportions of the target
variable based on the current weights.}
}
\value{
The calculated error, a single scalar greater than 0.
}
\description{
Returns the error between the current calculated weighted proportions and
the target proportions. Method of calculation depends on the error function
used. These are private function which is not intended for end-user use.
}
\details{
If supplying a custom error function closure to \code{\link{harvest}}, the
function should take two arguments: first, the target proportions, and second
the current weighted proportions based on the current provisional weights.

Error functions:
\describe{
  \item{\code{current_miss_linear}}{Sum of absolute differences}
  \item{\code{current_miss_squared}}{Sum of squared differences}
  \item{\code{current_miss_max}}{Maximum absolute difference}
  \item{\code{current_miss_mean}}{Mean absolute difference}
  \item{\code{current_miss_max_squared}}{Maximum squared difference}
  \item{\code{current_miss_mean_squared}}{Mean squared difference}
}
}
\keyword{internal}
