% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment_nas.R
\name{freeze_na_proportion}
\alias{freeze_na_proportion}
\title{Adjust target proportions to add a category for missing data}
\usage{
freeze_na_proportion(data, target, which_vars)
}
\arguments{
\item{data}{A data frame (tibble) or matrix containing sample data,
specified as in \code{\link{harvest}}}

\item{target}{A list of named vectors, specified as in \code{\link{harvest}}}

\item{which_vars}{A boolean, numeric, or character vector specifying which
variables to adjust in this manner. If boolean, all variables are selected.
If character, those named variables are selected. If numeric, variables
are selected based on numeric position amongst the target variables.}
}
\value{
An augmented target list, where affected variables with observed
  missing data in the sample has an additional level named "___NA"
  reflecting the proportion of data missing in the sample and other levels
  are adjusted accordingly
}
\description{
This function adds an additional level to any target variable to represent
missing data. The proportion given to the missing data is the proportion
missing in the observed sample. Other proportions are adjusted accordingly.
This is a private function and not intended for end-user use.
}
\keyword{internal}
