% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize}
\alias{normalize}
\title{Normalize weighting targets to sum to 1}
\usage{
normalize(target)
}
\arguments{
\item{target}{A list of target proportions in the population of interest.
This argument can be one of two formats: a list of named numeric vectors,
or a data frame (tibble). If a data frame, see the \code{target_map}
argument and the "Specifying \code{target}" of this documentation for
more details on the data frame's format. No level may have a negative
proportion or an NA, and each variable should sum to 1.}
}
\value{
A list of targets ready to be passed to
  \code{\link[harvest]{harvest}}
}
\description{
Normalize weighting targets to sum to 1
}
