% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mm.R
\name{plot_mm}
\alias{plot_mm}
\title{Many Models Evaluation Matrix}
\usage{
plot_mm(dataframe, sort)
}
\arguments{
\item{dataframe}{Dataframe to operate on. Must come from auto_mm}

\item{sort_values}{Which variable to sort models by?}
}
\description{
Plot from auto_mm
}
