% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{respondent_data}
\alias{respondent_data}
\title{Survey respondents}
\format{An object of class \code{data.frame} with 6703 rows and 18 columns.}
\usage{
respondent_data
}
\description{
A sample of 6,703 survey respondents from a single wave of the Nationscape
survey. Column names match \code{\link{ns_target}} in addition to a
single identifier \code{ResponseID} to track respondents. All variables are
character.
}
\details{
Respondents to the survey were permitted to leave
\code{household_income} blank, resulting in 652 missing values in this
sample.
}
\seealso{
\code{\link{ns_target}} data in this package.
}
\author{
Nationscape Team, University of California Los Angeles. Alex
Rossell-Hayes, Aaron Rudkin, Tyler Reny, Chris Tausanovitch, and Lynn
Vavreck.
}
\keyword{datasets}
