% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment_nas.R
\name{update_na_values}
\alias{update_na_values}
\title{Internal copy of data to overwrite missing data}
\usage{
update_na_values(data, target, which_vars)
}
\arguments{
\item{data}{A data frame (tibble) or matrix containing sample data,
specified as in \code{\link{harvest}}}

\item{target}{A list of named vectors, specified as in \code{\link{harvest}}}

\item{which_vars}{A boolean, numeric, or character vector specifying which
variables to adjust in this manner. If boolean, all variables are selected.
If character, those named variables are selected. If numeric, variables
are selected based on numeric position amongst the target variables.}
}
\value{
An augmented version of \code{data} with NAs of affected variables
  replaced with "___NA", and the columns subset to those in
  \code{target}
}
\description{
This function replaces NAs in any columns in order to allow for the strategy
described in \code{\link{freeze_na_proportion}}. NA values are replaced with
"___NA". No modifications done this way are saved and data returned by
\code{\link{harvest}} if \code{attach_weights=TRUE} will not be modified.
This is a private function and not intended for end-user use.
}
\keyword{internal}
