% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted_pct.R
\name{weighted_pct}
\alias{weighted_pct}
\title{Weighted proportion table}
\usage{
weighted_pct(x, weights)
}
\arguments{
\item{x}{A vector of any type}

\item{weights}{A numeric vector of weights, equal in length to \code{x}}
}
\value{
A named vector with weighted proportions.
}
\description{
This is a fast, dependency-free weighted proportion table for 1D variables.
}
\examples{
# Setup examle data
x <- c(rep(1, 5), rep(2, 5))
weights <- c(rep(1, 5), rep(2, 5))

# Unweighted proportions: 0.5 / 0.5
prop.table(table(x))
# Weighted proportions: 0.33 / 0.66
weighted_pct(x, weights)
}
