% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_effect.R
\name{design_effect}
\alias{design_effect}
\title{Calculate design effect from weights}
\usage{
design_effect(weights, outcome = NULL)
}
\arguments{
\item{weights}{The weights for which a design effect is desired}

\item{outcome}{A vector outcome of interest.}
}
\value{
A numeric design effect for the weights
}
\description{
This function calculates the design effect of the supplied weights. A design
effect is an estimate of the variance inflation caused by the design of a
study. If a vector of weights is provided without any outcome, this function
implements the estimator in Kish (1992), which assumes no correlation between
provided weights and an outcome of interest. If an outcome of interest is
provided, this function implements the estimator in Spencer (2000), which
accounts for correlation between outcomes and weights. For a broader
discussion of design effects in weighted surveys, see Spencer (2000).
}
