% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rake.R
\name{do_rake}
\alias{do_rake}
\title{Performs iterative raking on data}
\usage{
do_rake(data, target, weights, max_weight, max_iterations, convergence,
  enforce_mean = TRUE, verbose)
}
\arguments{
\item{data}{The data frame (tibble) or matrix containing the original data}

\item{target}{The target proportions, specified as in \code{\link{harvest}}}

\item{weights}{A numeric vector of current weights}

\item{max_weight}{The maximum weight to clamp weights to after raking each
variable}

\item{max_iterations}{The maximum number of iterations of raking to perform
before giving up. Please note that depending on the variable selection
method, \code{\link{harvest}} may continue iterative raking on new
variables after completing the iterations of this function call.}

\item{convergence}{Convergence parameters as described in
\code{\link{harvest}}.}

\item{enforce_mean}{Default TRUE. By default, weights minimize divergence
from target proportions subject to two conditions: that the mean weight
be 1, and that the maximum weight be capped at \code{max_weight}. Weights
are first capped and then re-meaned. When \code{enforce_mean} is FALSE,
the re-meaning does not occur. This will guarantee the maximum weight
does not exceed \code{max_weight} but may result in mean weights diverging
from 1. As \code{max_weight} prevents high-weight observations from
becoming higher weighted, \code{enforce_mean} helps low-weight observations
from becoming even lower weighted.}

\item{verbose}{Level of verbosity, defaults to FALSE. At TRUE or 1, the
function begins emitting progress information during major events. At 2,
each iteration provides progress information.}
}
\value{
A numeric vector of adjusted, raked weights.
}
\description{
This function performs iterative raking as described in DeBell and Krosnick
(2009). It is a private function and not intended for end-users. End-users
should call \code{\link{harvest}} to add weights.
}
\keyword{internal}
