% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rake.R
\name{single_adjust}
\alias{single_adjust}
\title{Adjust weights based on current imbalance of a single variable.}
\usage{
single_adjust(data, weights, target, var, cache)
}
\arguments{
\item{data}{The data frame (tibble) or matrix which contains the original
data.}

\item{weights}{The current weights}

\item{target}{The target proportions, specified as in \code{\link{harvest}}}

\item{var}{A quoted character vector containing the variable we are currently
raking on.}

\item{cache}{A list of integer vectors mapping rows of the data frame to
levels of the target variable}
}
\value{
Adjusted weights, raked for this variable.
}
\description{
This function internally rakes the weights provided according to the
current imbalance in a single variable. The mathematical process here is
simple: for each level of the variable, multiply the weight by the target
proportion divided by the current weighted proportion. For instance, if a
\code{party} variable should be 0.4 Independent, 0.35 Democratic, and 0.25
Republican, and the current weighted proportion is 0.4 Democratic, then the
weights of every Democrat will be multiplied by 0.35 / 0.4. This is an
internal function and not intended for end-users.
}
\keyword{internal}
